Add plotting of avg time, add one more dimension and probability

+++ experiment.py
@@ -5,41 +5,58 @@ import os
 import pickle
 import time
 import pandas as pd
+import matplotlib.pyplot as plt
 
-# TODO fare un grafico con ascisse: tempo ordinate: probabilità
-def evaluate_experiment(dim):
+def evaluate_experiment(probabilities, dim):
     df = pd.DataFrame(columns= ['Probabilities', 'MST_time'])
-    df['Probabilities'] = [0.5, 0.75, 0.85, 0.9]
+    df['Probabilities'] = probabilities
     graph = Graph.Graph()
+    m_elapsed = []
+    elapsed = []
     file_list = []
     file_name = 'graph_' + str(dim) + '_'
     for files in os.listdir('db'):
         if file_name in files:
             file_list.append(files)
-    for files, prob in zip(file_list, df['Probabilities']):
-        name = 'db/' + files
-        print(name)
-        file = open(name, 'rb')
-        graph = pickle.load(file)  # load file and calculate MST
-        file.close()
-        start = time.perf_counter()
-        MST_Kruskal.MST_Kruskal(graph)
-        end = time.perf_counter()
-        df['MST_time'] = end - start
+    for files in file_list:
+        for i in range(1, 11):
+            name = 'db/' + files
+            print(name)
+            file = open(name, 'rb')
+            graph = pickle.load(file)  # load file and calculate MST
+            file.close()
+            start = time.perf_counter()
+            MST_Kruskal.MST_Kruskal(graph)
+            end = time.perf_counter()
+            elapsed.append(end - start)
+        m_elapsed.append(sum(elapsed)/10)
+    df['MST_time'] = m_elapsed
     print(df)
-
+    plt.xlabel("Probability")
+    plt.ylabel("Time to find MST")
+    plt.plot(df['Probabilities'], df['MST_time'], marker="o")
+    plt.title("MST benchmark on graph with " + str(dim) + " nodes")
+    try:
+        os.mkdir("figures")
+    except FileExistsError:
+        pass
+    plt.savefig('figures/' + file_name + '.png')
+    plt.clf()
 
 
 # this function generate random graphs, takes in input probability 
-def generate_random_graph(prob):
-    n_nodes = [50, 100, 500, 1000]
+def generate_random_graph(prob, dimensions):
+    n_nodes = dimensions
     for node in n_nodes:
-        graph = Graph.Graph(node)
-        graph.random_weighted_graph(prob)
         file_name = "db/graph_" + str(node) + "_prob_" + str(prob) + ".txt"
-        file = open(file_name, 'wb')
-        pickle.dump(graph, file)
-        file.close()
+        if os.path.isfile(file_name): # if exist don't create new files
+            continue
+        else:
+            graph = Graph.Graph(node)
+            graph.random_weighted_graph(prob)
+            file = open(file_name, 'wb')
+            pickle.dump(graph, file)
+            file.close()
     
 
 
@@ -49,6 +66,10 @@ def start_experiment():
     except FileExistsError:
         print("Directory already exist")
         pass
-    probabilities = [0.5, 0.75, 0.85, 0.9] # list of probabilities
+    probabilities = [0.25, 0.5, 0.75, 0.85, 0.9] # list of probabilities
+    dimensions = [10, 50, 100, 500, 1000]
     for prob in probabilities:
-        generate_random_graph(prob)
+        generate_random_graph(prob, dimensions)
+    for dim in dimensions:
+        evaluate_experiment(probabilities, dim)
+
